replace utf8-parsing code with utf8 crate
The Read::chars method is gone for good; no point in waiting for it. Had
to tweak the error messages a bit because `utf8` provides slightly
different results for the bad bytes when an error is occurred.

Also changed the behavior when bad UTF-8 is encountered. Previously,
this was a fatal error. Now, an error is printed, but the input is
replaced with U+FFFE (replacement character) and processing continues.