fix some EOF-related bugs in parser
Remove the janky ParserState::Unused; we can just use the "did the
Option get take'n or not" like Parser::next <-> Parser::Step does. This
removes an unnecessary layer of char stashing and makes it easier to see
what's going on.

When fixing this ugliness, some bugs in EOF handling were found, so
those have been fixed and tests added (and some tests added for things
that weren't bugs but could have been).