make the parser free of any buffering
The parser refactor merged in v2 introduced a perf regression compared
to v1, which I've been able to narrow down to increased String
allocations.

This is due to the fact that the parser buffers the values of strings
and numbers internally before handing over that owned buffer to the DC4
state machine.

For strings this is okay because that buffer just goes straight into the
stack, but for numbers it's thrown away after the number characters are
parsed into a BigReal (which only the state machine can do because it
depends on values like iradix and scale).

The solution is to change the parser to not buffer these, and split the
Actions for strings and numbers into two actions each: one for each
character, finished with an action to push them to the stack.