operate on bytes, not UTF-8 strings
DC4 has always treated its input as UTF-8, but I think that this was not
actually the correct stance to take. The commands are all ASCII
characters, the number of registers is defined as 256, and things like
the 'a' command interpret strings as bytes. It simplifies things quite a
bit if we just treat the input as arbitrary bytes, and doesn't affect
correctness at all because that's what traditional Unix dc does.