jit: remember the parsed actions for string values
This change adds a field to string values that holds a ref-counted vec
of the parsed actions that the string represents, which is filled in
lazily when the string is first executed as a macro. Subsequent
executions avoid having to run the parser again. For programs with a lot
of looping, this can result in a dramatic speed-up.